chore: update deployment scripts to reflect new architecture that includes SignedMultiCallDeploy contract
