+++ scripts/v2/step0.json
@@ -1,0 +1,6 @@
+{
+ "signerRootAdminPubKey": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+ "signerAdminPubKey": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+ "signerAddress": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+ "immutableSigner": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
+}
