+++ scripts/v2/step2.json
@@ -1,5 +1,5 @@
 {
- "walletImplLocatorAdmin": "0xccBd3382EA55CF431126Ac5c4E289E100040aa2B",
- "walletImplChangerAdmin": "0xccBd3382EA55CF431126Ac5c4E289E100040aa2B",
- "latestWalletImplLocator": "0x8535282CF49A5B6326a19f0D190DcC3f2028B53a"
+ "walletImplLocatorAdmin": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+ "walletImplChangerAdmin": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+ "latestWalletImplLocator": "0x7A02b0bC209F19Fe0b2d06cAb748aDb82844e3A2"
 }
