+++ scripts/v2/step3.json
@@ -1,4 +1,4 @@
 {
- "walletImplLocatorAddress": "0x8535282CF49A5B6326a19f0D190DcC3f2028B53a",
- "startupWalletImpl": "0x0d1Bf2f4Ab334324665aeb8d481fF92CdE289439"
+ "walletImplLocatorAddress": "0x7A02b0bC209F19Fe0b2d06cAb748aDb82844e3A2",
+ "startupWalletImpl": "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853"
 }
