--- scripts/v2/step5.json
@@ -1,6 +1,0 @@
-{
- "signerRootAdminPubKey": "0xccBd3382EA55CF431126Ac5c4E289E100040aa2B",
- "signerAdminPubKey": "0xccBd3382EA55CF431126Ac5c4E289E100040aa2B",
- "signerAddress": "0xccBd3382EA55CF431126Ac5c4E289E100040aa2B",
- "immutableSigner": "0xb679fECE82851b3082204F08c376fC0a67c57AB7"
-}
