+++ scripts/v2/step6.json
@@ -1,0 +1,3 @@
+{
+ "nexusBootstrap": "0x610178dA211FEF7D417bC0e6FeD39F05609AD788"
+}
