--- scripts/v2/step8.json
@@ -1,4 +1,0 @@
-{
- "mockValidator": "0x48a515007c7bb6496cdf82B857B0dBC21579C541",
- "mockExecutor": "0x4811EB68188Db64148989fd72F3e819a42C51dcd"
-}
